--- evm_bridge/src/serde/serde.ts
@@ -1,7 +1,0 @@
-import { Scalar } from "o1js";
-
-export namespace SerdeJSON {
-    export function deserScalars(strs: string[]): Scalar[] {
-        return strs.map(Scalar.from);
-    }
-}
