+++ evm_bridge/src/verifier/batching.test.ts
@@ -3,12 +3,10 @@ import { PolyComm } from "../poly_commitment/commitment";
 import { SRS } from "../SRS";
 import { Batch } from "./batching";
 import { VerifierIndex } from "./verifier";
-import proof_evals from "../../test/proof_evals.json" assert { type: "json" };
+import proof_evals_json from "../../test/proof_evals.json" assert { type: "json" };
 import { PointEvaluations } from "../prover/prover";
+import { deserProofEvals } from "../serde/serde_proof";
 
-function proofEvalsFromJSON(evals: ProofEvalsJSON): {
-
-}
 
 test("toBatch() step 1", () => {
     const srs = SRS.createFromJSON();
@@ -18,6 +16,7 @@ test("toBatch() step 1", () => {
         domain_size: domain_size,
         public: 0
     };
+    const proof_evals = deserProofEvals(proof_evals_json);
     const proof = {
         evals: proof_evals
     };
